+++ lib/http_stub/server/application/cross_origin_support.rb
@@ -9,14 +9,24 @@ module HttpStub
           def add_headers_if_necessary
             response.headers.merge!(
               "Access-Control-Allow-Origin"  => "*",
-              "Access-Control-Allow-Methods" => @http_stub_request.method.upcase,
-              "Access-Control-Allow-Headers" => @http_stub_request.headers.keys.join(",")
+              "Access-Control-Allow-Methods" => determine_allowed_methods,
+              "Access-Control-Allow-Headers" => determine_allowed_headers
             ) if settings.cross_origin_support?
           end
 
           def handle_options_if_necessary
             pass unless settings.cross_origin_support?
             halt 200
+          end
+
+          private
+
+          def determine_allowed_methods
+            @http_stub_request.headers["Access_Control_Request_Method"] || @http_stub_request.method.upcase
+          end
+
+          def determine_allowed_headers
+            @http_stub_request.headers["Access_Control_Request_Headers"] || @http_stub_request.headers.keys.join(",")
           end
 
         end